/save_route update request; refactor db_endpoints organization
