+++ models.py
@@ -53,6 +53,11 @@ class AddRequestInput(BaseModel):
     coordinates: List[Point]
 
 
+class SaveRouteInput(BaseModel):
+    request_id: int
+    points: TSPinput
+
+
 class RouteInfo(BaseModel):
     route_id: int
